--- tests/models/steps/processing/dataset_version_creation/conftest.py
@@ -1,3 +1,0 @@
-pytest_plugins = [
-    "tests.models.steps.processing.dataset_version_creation.fixtures.diversified_data_extractor_processing_fixtures"
-]
