+++ tests/models/steps/processing/dataset_version_creation/fixtures/diversified_data_extractor_processing_fixtures.py
@@ -3,24 +3,12 @@ import pytest
 import open_clip
 
 
-from picsellia import Client, Datalake
-
 from src.models.steps.processing.dataset_version_creation.diversified_data_extractor_processing import (
     DiversifiedDataExtractorProcessing,
 )
 from src.steps.model_loader.processing.processing_diversified_data_extractor_model_loader import (
     OpenClipEmbeddingModel,
 )
-
-
-@pytest.fixture
-def picsellia_client(api_token: str, host: str) -> Client:
-    return Client(api_token=api_token, host=host)
-
-
-@pytest.fixture
-def picsellia_datalake(picsellia_client) -> Datalake:
-    return picsellia_client.get_datalake()
 
 
 @pytest.fixture(scope="session")
@@ -51,14 +39,14 @@ def mock_open_clip_embedding_model(mock_model_architecture, mock_pretrained_weig
 @pytest.fixture
 def diversified_data_extractor_processing(
     picsellia_client,
-    picsellia_datalake,
+    picsellia_default_datalake,
     mock_dataset_context,
     mock_dataset_version,
     mock_open_clip_embedding_model,
 ):
     return DiversifiedDataExtractorProcessing(
         client=picsellia_client,
-        datalake=picsellia_datalake,
+        datalake=picsellia_default_datalake,
         input_dataset_context=mock_dataset_context,
         output_dataset_version=mock_dataset_version,
         embedding_model=mock_open_clip_embedding_model,
